# packages/api-spec/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/printers": {
      "get": {
        "operationId": "PrintersController_list",
        "summary": "List all printers",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of printers",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PrinterResponse"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "printers"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "PrintersController_create",
        "summary": "Create a printer",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePrinterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created printer",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrinterResponse"
                }
              }
            }
          }
        },
        "tags": [
          "printers"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/printers/{id}": {
      "get": {
        "operationId": "PrintersController_get",
        "summary": "Get printer by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Printer details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrinterResponse"
                }
              }
            }
          }
        },
        "tags": [
          "printers"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "operationId": "PrintersController_update",
        "summary": "Update a printer",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePrinterDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated printer",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrinterResponse"
                }
              }
            }
          }
        },
        "tags": [
          "printers"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/printers/{id}/status": {
      "get": {
        "operationId": "PrintersController_checkStatus",
        "summary": "Check printer TCP reachability",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Printer reachability status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrinterStatusResponse"
                }
              }
            }
          }
        },
        "tags": [
          "printers"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/printers/{id}/test": {
      "post": {
        "operationId": "PrintersController_testPrint",
        "summary": "Print a test ticket",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Test ticket sent",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        },
        "tags": [
          "printers"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/reports/x": {
      "get": {
        "operationId": "ReportsController_getXReport",
        "summary": "Live X-report for the current open day",
        "parameters": [],
        "responses": {
          "200": {
            "description": "X-report snapshot"
          }
        },
        "tags": [
          "reports"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/reports/z/{dayId}": {
      "get": {
        "operationId": "ReportsController_getZReport",
        "summary": "Z-report for a closed day",
        "parameters": [
          {
            "name": "dayId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Z-report detail"
          }
        },
        "tags": [
          "reports"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/reports/sales": {
      "get": {
        "operationId": "ReportsController_getSales",
        "summary": "Daily sales totals over a date range",
        "parameters": [
          {
            "name": "from",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Daily sales totals"
          }
        },
        "tags": [
          "reports"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/reports/vat": {
      "get": {
        "operationId": "ReportsController_getVat",
        "summary": "VAT summary over a date range (for VAT return)",
        "parameters": [
          {
            "name": "from",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "VAT summary with grand total"
          }
        },
        "tags": [
          "reports"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/reports/z/{dayId}/print": {
      "post": {
        "operationId": "ReportsController_printZReport",
        "summary": "Print Z-report on receipt printer",
        "parameters": [
          {
            "name": "dayId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Print result"
          }
        },
        "tags": [
          "reports"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/reports/x/print": {
      "post": {
        "operationId": "ReportsController_printXReport",
        "summary": "Print X-report on receipt printer",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Print result"
          }
        },
        "tags": [
          "reports"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "SpicyHome POS API",
    "description": "REST API for the SpicyHome restaurant POS system",
    "version": "0.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "PrinterResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Kitchen"
          },
          "ip": {
            "type": "string",
            "example": "192.168.1.100"
          },
          "port": {
            "type": "integer",
            "format": "int32",
            "example": 9100
          },
          "role": {
            "type": "string",
            "example": "kitchen"
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "createdAt": {
            "type": "integer",
            "format": "int64",
            "example": 1700000000
          },
          "updatedAt": {
            "type": "integer",
            "format": "int64",
            "example": 1700000000
          },
          "createdBy": {
            "type": "integer",
            "format": "int64",
            "example": 1,
            "nullable": true
          },
          "updatedBy": {
            "type": "integer",
            "format": "int64",
            "example": 1,
            "nullable": true
          }
        },
        "required": [
          "id",
          "name",
          "ip",
          "port",
          "role",
          "isActive",
          "createdAt",
          "updatedAt",
          "createdBy",
          "updatedBy"
        ]
      },
      "CreatePrinterDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Kitchen"
          },
          "ip": {
            "type": "string",
            "example": "192.168.1.100"
          },
          "port": {
            "type": "integer",
            "format": "int32",
            "default": 9100
          },
          "role": {
            "type": "string",
            "enum": [
              "receipt",
              "kitchen"
            ],
            "example": "kitchen"
          },
          "isActive": {
            "type": "boolean",
            "default": true
          }
        },
        "required": [
          "name",
          "ip",
          "role"
        ]
      },
      "UpdatePrinterDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "ip": {
            "type": "string"
          },
          "port": {
            "type": "integer",
            "format": "int32"
          },
          "role": {
            "type": "string",
            "enum": [
              "receipt",
              "kitchen"
            ]
          },
          "isActive": {
            "type": "boolean"
          }
        }
      },
      "PrinterStatusResponse": {
        "type": "object",
        "properties": {
          "reachable": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "reachable"
        ]
      },
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "success"
        ]
      },
...
    }
  }
}